# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/gradebooks": {
      "get": {
        "operationId": "gradebooksList",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookListDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [] }]
      },
      "post": {
        "operationId": "gradebooksCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateGradebookInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/cells": {
      "get": {
        "operationId": "gradebooksCells",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CellsResponseDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/grades": {
      "put": {
        "operationId": "gradebooksBatchUpdate",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchUpdateInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchUpdateResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/components/{componentId}/deadline": {
      "put": {
        "operationId": "gradebooksSetDeadline",
        "parameters": [
          {
            "name": "componentId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GradeDeadlineInput" }
            }
          }
        },
        "responses": {
          "204": { "description": "" },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/components/{componentId}/lock": {
      "post": {
        "operationId": "gradebooksLockColumn",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "componentId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LockInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/lock": {
      "post": {
        "operationId": "gradebooksLock",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LockInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/sync-roster": {
      "post": {
        "operationId": "gradebooksSyncRoster",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LockInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [], "csrf": [] }]
      }
    },
    "/api/v1/gradebooks/{id}/history": {
      "get": {
        "operationId": "gradebooksHistoryAll",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GradebookHistoryDto" }
              }
            }
          },
          "400": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "429": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          },
          "500": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorDto" }
              }
            }
          }
        },
        "tags": ["gradebooks"],
        "security": [{ "bearer": [] }, { "cookie": [] }]
      }
    },
...
  },
  "info": {
    "title": "Quản lý điểm API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "cookie": { "type": "apiKey", "in": "cookie", "name": "qld_session" },
      "csrf": { "type": "apiKey", "in": "header", "name": "x-csrf-token" }
    },
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "object", "nullable": true },
          "requestId": { "type": "string" }
        },
        "required": ["code", "message", "details", "requestId"]
      },
...
      "GradebookDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "classId": { "type": "number" },
          "subjectId": { "type": "number" },
          "termId": { "type": "number" },
          "className": { "type": "string" },
          "subjectName": { "type": "string" },
          "termName": { "type": "string" },
          "status": { "type": "string", "enum": ["DANG_NHAP_LIEU", "DA_CHOT"] },
          "version": { "type": "number" }
        },
        "required": [
          "id",
          "classId",
          "subjectId",
          "termId",
          "className",
          "subjectName",
          "termName",
          "status",
          "version"
        ]
      },
      "GradebookListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GradebookDto" }
          },
          "nextCursor": { "type": "number", "nullable": true }
        },
        "required": ["items", "nextCursor"]
      },
      "CreateGradebookInput": {
        "type": "object",
        "properties": {
          "classId": { "type": "number" },
          "subjectId": { "type": "number" },
          "termId": { "type": "number" }
        },
        "required": ["classId", "subjectId", "termId"]
      },
...
      "CellsResponseDto": {
        "type": "object",
        "properties": {
          "book": { "$ref": "#/components/schemas/GradebookDto" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GradeCellDto" }
          },
          "nextCursor": { "type": "string", "nullable": true }
        },
        "required": ["book", "items", "nextCursor"]
      },
...
      "LockInput": {
        "type": "object",
        "properties": { "expectedVersion": { "type": "number" } },
        "required": ["expectedVersion"]
      },
...
    }
  }
}